% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafonction.R
\name{mafonction}
\alias{mafonction}
\title{Une fonction pour l'exemple}
\usage{
mafonction(nom = "Wei")
}
\arguments{
\item{nom}{Saisi par utilisateur}
}
\value{
Chaine de caractere salut XXX !
}
\description{
Cette fonction sert a montrer le fonctionnement de procedure.
}
\examples{
mafonction() 
}
\keyword{salutation}
